DtoMapper update

+++ src/main/java/pl/tkaras/carworkshopwebservice/logic/CarService.java
@@ -1,6 +1,9 @@
 package pl.tkaras.carworkshopwebservice.logic;
 
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import pl.tkaras.carworkshopwebservice.model.dto.CarDto;
+import pl.tkaras.carworkshopwebservice.model.mapper.impl.CarDtoMapper;
 import pl.tkaras.carworkshopwebservice.model.entity.Car;
 import pl.tkaras.carworkshopwebservice.repository.CarRepository;
 
@@ -10,25 +13,35 @@ import java.util.List;
 public class CarService {
 
     private CarRepository carRepo;
+    private CarDtoMapper carDtoMapper;
 
     private CarService(CarRepository carRepo) {
         this.carRepo = carRepo;
     }
 
-    public Car getSingleCar(Long id){
-        return carRepo.findById(id).orElseThrow();
+    public CarDto getSingleCar(Long id){
+        if(!carRepo.existsById(id)){
+            //return ResponseEntity.notFound().build();
+        }
+        return new CarDtoMapper().mapToDto(carRepo.findById(id).orElseThrow());
     }
 
-    public List<Car> gelAllCars(){
-        return carRepo.findAll();
+    public List<CarDto> gelAllCars(){
+        return carDtoMapper.mapToDtos(carRepo.findAll());
     }
 
-    public Car addCar(Car entity){
-        return carRepo.save(entity);
+    public CarDto addCar(Car entity){
+        return carDtoMapper.mapToDto(carRepo.save(entity));
     }
 
-    public void deleteCar(Long id){
-        carRepo.deleteById(id);
+    public ResponseEntity deleteCar(Long id){
+        if(carRepo.existsById(id)){
+            carRepo.deleteById(id);
+        }
+        else{
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok().build();
     }
